pyproject.toml
[project]
name = "exambot-assessments-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb[duckdb]>=1.0.11",
    "langgraph>=0.4.7",
    "pandas>=2.2.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
]
